# Conflicts:
#	tools/ci_build/github/linux/docker/inference/aarch64/default/cpu/scripts/install_centos.sh
#	tools/ci_build/github/linux/docker/inference/x64/default/cpu/scripts/install_centos.sh
#	tools/ci_build/github/linux/docker/inference/x64/python/cpu/scripts/install_centos.sh

+++ onnxruntime/core/optimizer/qdq_transformer/avx2_weight_s8_to_u8.cc
@@ -99,7 +99,7 @@ static bool TryConvertDynamicQuantizeLSTM(Node& op_node, Graph& graph) {
   Initializer w_temp(*weight_tensor_proto, graph.ModelPath());
   {
     int8_t* p = w_temp.data<int8_t>();
-    for (int i = 0; i < w_temp.size(); i++) {
+    for (size_t i = 0; i < w_temp.size(); i++) {
       if (*p < -64 || *p > 64) {
         should_convert = true;
       }
@@ -111,7 +111,7 @@ static bool TryConvertDynamicQuantizeLSTM(Node& op_node, Graph& graph) {
   Initializer r_temp(*r_tensor_proto, graph.ModelPath());
   {
     int8_t* p = r_temp.data<int8_t>();
-    for (int i = 0; i < r_temp.size(); i++) {
+    for (size_t i = 0; i < r_temp.size(); i++) {
       if (*p < -64 || *p > 64) {
         should_convert = true;
       }
@@ -129,7 +129,7 @@ static bool TryConvertDynamicQuantizeLSTM(Node& op_node, Graph& graph) {
   weights_proto_u8.set_data_type(ONNX_NAMESPACE::TensorProto_DataType_UINT8);
   weights_proto_u8.set_name(weight_tensor_proto->name() + "_s8_2_u8");
   weights_proto_u8.mutable_dims()->CopyFrom(weight_tensor_proto->dims());
-  weights_proto_u8.set_raw_data(w_temp.data<int8_t>(), w_temp.size());
+  weights_proto_u8.set_raw_data(w_temp.data<int8_t>(), static_cast<size_t>(w_temp.size()));
   input_defs[w_idx] = &graph_utils::AddInitializer(graph, weights_proto_u8);
 
   ONNX_NAMESPACE::TensorProto weight_zp_proto_u8;
@@ -140,7 +140,7 @@ static bool TryConvertDynamicQuantizeLSTM(Node& op_node, Graph& graph) {
   r_proto_u8.set_data_type(ONNX_NAMESPACE::TensorProto_DataType_UINT8);
   r_proto_u8.set_name(r_tensor_proto->name() + "_s8_2_u8");
   r_proto_u8.mutable_dims()->CopyFrom(r_tensor_proto->dims());
-  r_proto_u8.set_raw_data(r_temp.data<int8_t>(), r_temp.size());
+  r_proto_u8.set_raw_data(r_temp.data<int8_t>(), static_cast<size_t>(r_temp.size()));
   input_defs[r_idx] = &graph_utils::AddInitializer(graph, r_proto_u8);
 
   ONNX_NAMESPACE::TensorProto r_zp_proto_u8;
